Split reverse dependencies into Jane Street and non Jane Street
The Jane Street packages on this universe are checked out to v0.14
since v0.15 wasn't on opam yet when creating the universe. So building
the Jane Street packages on this universe isn't that helpful. That's
why this commit adds an option to the `build` helper to only build
the non Jane Street packages.